refactor: move all home sections closer to the Home page
