--- web/src/pages/Home.tsx
@@ -1,19 +1,0 @@
-import TestimonialsSection from '../components/testimonials/TestimonialsSection';
-import BenefitsSection from '../components/benefits/BenefitsSection';
-import AboutSection from '../components/about/AboutSection';
-import HeroSection from '../components/hero/HeroSection';
-import NewsSection from '../components/news/NewsSection';
-
-function HomePage() {
-  return (
-    <>
-      <HeroSection />
-      <AboutSection />
-      <TestimonialsSection />
-      <BenefitsSection />
-      <NewsSection />
-    </>
-  );
-}
-
-export default HomePage;
