+++ web/src/pages/Home/components/Sections/hero/HeroText.tsx
