+++ web/src/pages/Home/index.tsx
@@ -1,0 +1,17 @@
+import TestimonialsSection from './components/Sections/testimonials/TestimonialsSection';
+import BenefitsSection from './components/Sections/benefits/BenefitsSection';
+import AboutSection from './components/Sections/about/AboutSection';
+import HeroSection from './components/Sections/hero/HeroSection';
+import NewsSection from './components/Sections/news/NewsSection';
+
+export default function HomePage() {
+  return (
+    <>
+      <HeroSection />
+      <AboutSection />
+      <TestimonialsSection />
+      <BenefitsSection />
+      <NewsSection />
+    </>
+  );
+}
